feat: MyBookDetailTypeMap を作成
